+++ lib/src/utils/create_function_context.dart
@@ -1,7 +1,5 @@
 import 'dart:convert';
 
-import 'package:synapp_functions_helper/src/function_wrapper.dart';
-import 'package:synapp_functions_helper/src/models/function_context.dart';
 import 'package:synapp_functions_helper/synapp_functions_helper.dart';
 
 FunctionContext createFunctionContext(