+++ CHANGELOG.md
@@ -1,3 +1,20 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.0.1
+
+- add ignored files
+
+## 1.0.2
+
+- functionWrapper returns future
+
+
+## 1.0.3
+
+- versions for sdk min 17
+
+## 1.0.4
+
+- await in function wrapper exception